requirement.txt
langchain_core
python-dotenv
streamlit
langchain_community
langchain-ollama
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
chromadb
beautifulsoup4
sentence_transformers
PyPDF2
python-dotenv
